Added lines to use ROOT_CINT in requirements
Lots of additions.  Essentially adding framework for channel by channel calibrations of ACD.

AcdMap is just a bunch of ids and mappings.  Name may have to change to avoid conflicts w/ other packages.

AcdCalibResult, AcdGainFitResult and AcdPedFitResult are storage for calibration results on a single channel.

AcdCalibMap, AcdGainCalibMap and AcdPedestalCalibMap are maps from channel ID -> result objects.

AcdHistCalibMap is a map from channel ID -> histogram

Acd*Library are the libraries of fitting functions.

AcdCalibBase is the base class for the calibration drivers.

acdSel and muonCalib are calibration drivers for offline ACD data and muon data respecteively.